Revert "Allow geometries to extend a bit more past their prospective parents (#32)"
This reverts commit d5157bc407180742a09da7a7a44eadd1bc4167ab.

Not sure exactly why, but the value used in the reverted commit was way
off (area difference is compared to 300 while the max value encountered
in the whole import seems to be way under 1) causing the parent
matching to fail because of multiple matching parents. Also, using the
previous value 1e-6 seems to yield correct results.