Use gettext_lazy in models.py
Using regular gettext leads to a possible localization clash in apps
which have munigeo as a dependency. For example running makemigrations
on such an app that has a translation defined for "Name" results in a
new migration being generated for munigeo. This in turn is undesirable
for CI pipelines which would like to check migrations.